requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
Django==2.2.16
djangorestframework==3.12.4
idna==3.4
iniconfig==1.1.1
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
requests==2.26.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.13
